+++ linode/instancenetworking/schema_datasource.go
@@ -1,4 +1,4 @@
-package instanceip
+package instancenetworking
 
 import "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
